Prevent Sphinx from resolving paramlinks as attrs
The attempt to add new attributes Column.index / Column.unique
revealed that Sphinx would resolve a paramlink if the
module/class/name looks like a class attribute, preventing
our resolver from working.

After attempts to work with ".param." as a first class prefix,
it was realized (as it likely was years ago) that this token
interferes with Sphinx "~" operator, so instead we now apply
an obfuscation token to the param name up front, preventing
Sphinx resolver from ever trying to resolve a paramlink.
We then remove the obfuscation in our own resolver where thankfully
we have access to all the text nodes that will be rendered.